Add double gatsby build hopefully build minutes don't become a problem
The idea here is that build:parcel throws if it's before gatsby-build
due to it using the same babel file, but we need to build the gatsby
files once the static files have been output by build:parcel so we have
to run gatsby build twice for a quick fix.

A possible other solution is to not rely on the same babel config for
both of these things and instead just different ones.